various changes, temporary
