+++ packages/shared/src/raven/index.js
@@ -2,7 +2,7 @@ require('dotenv').config({ path: '../' })
 const debug = require('debug')('shared:raven')
 // import config from '../config'
 const { NODE_ENV, FORCE_DEV, SENTRY_DSN_SERVER } = process.env
-console.log(SENTRY_DSN_SERVER)
+
 let Raven
 
 const options = {
